add extend.js
